+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+## 3.3.0 - 2024-05-24
+
 - feat: allow anonymous uuid generation to be customizable ([#132](https://github.com/PostHog/posthog-android/pull/132))
 - recording: fix removing the ViewTreeObserver when its not alive ([#134](https://github.com/PostHog/posthog-android/pull/134))